replace direct core lib usage with SciChartReact in LogarithmicAxis
